+++ src/external/rprand.h
@@ -186,7 +186,7 @@ int *rprand_load_sequence(unsigned int count, int min, int max)
 {
     int *sequence = NULL;
     
-    if (count > (abs(max - min) + 1)) 
+    if (count > (unsigned int)(abs(max - min) + 1)) 
     {
         RPRAND_LOG("WARNING: Sequence count required is greater than range provided\n");
         //count = (max - min);
@@ -198,9 +198,9 @@ int *rprand_load_sequence(unsigned int count, int min, int max)
     int value = 0;
     bool value_is_dup = false;
 
-    for (int i = 0; i < count;)
+    for (unsigned int i = 0; i < count;)
     {
-        value = (rprand_xoshiro()%(abs(max - min) + 1)) + min;
+        value = ((int)rprand_xoshiro()%(abs(max - min) + 1)) + min;
         value_is_dup = false;
 
         for (int j = 0; j < i; j++)
